% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ronants_core.R
\name{ronats_plot}
\alias{ronats_plot}
\title{Covid Tracking Project Time Series Plots}
\usage{
ronats_plot(d, v, ...)
}
\arguments{
\item{d}{A data frame with a date column named "date." Intended for use with
Covid Tracking Project data via ronats::ronats_pull()}

\item{v}{A non-date continuous variable. Intended for use with Covid Tracking
Project variables. See https://covidtracking.com/api for current and
deprecated variables available via the project API.}

\item{...}{Argument passed to other methods.}
}
\description{
This function returns a ggplot2 object plotting a Covid Tracking Project
  variable with the daily value as a column chart and 7-day, right-aligned
  moving averages and LOESS-smoothed trends as line graphs.
}
\examples{
nc <- ronats_pull(state = "nc")
nc_hosp <- ronats_plot(nc, hospitalizedCurrently)
}
