% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ronants_core.R
\name{ronats_pull}
\alias{ronats_pull}
\title{Pull data from Covid Tracking Project API}
\usage{
ronats_pull(state = "us")
}
\arguments{
\item{state}{If NULL (the default), pull data for all US states plus DC.
Otherwise, a 2-character lowercase string with the state's abbreviation,
e.g. "nc".}
}
\description{
Pull data from Covid Tracking Project API
}
\examples{
us <- ronats_pull()
head(us)

nc <- ronats_pull(state = "nc")
head(nc)

}
